+++ files/gitlab-cookbooks/gitlab-ee/recipes/pgbouncer.rb
@@ -44,6 +44,7 @@ end
 runit_service 'pgbouncer' do
   options(
     username: node['gitlab']['postgresql']['username'],
+    groupname: node['gitlab']['postgresql']['group'],
     data_directory: node['gitlab']['pgbouncer']['data_directory'],
     log_directory: node['gitlab']['pgbouncer']['log_directory']
   )