+++ files/gitlab-cookbooks/gitlab/libraries/account_helper.rb
@@ -43,7 +43,7 @@ class AccountHelper
   end
 
   def redis_group
-    node['gitlab']['redis']['username']
+    node['gitlab']['redis']['group']
   end
 
   def postgresql_user
@@ -51,7 +51,7 @@ class AccountHelper
   end
 
   def postgresql_group
-    node['gitlab']['postgresql']['username']
+    node['gitlab']['postgresql']['group']
   end
 
   def mattermost_user
@@ -72,6 +72,10 @@ class AccountHelper
 
   def prometheus_user
     node['gitlab']['prometheus']['username']
+  end
+
+  def prometheus_group
+    node['gitlab']['prometheus']['group']
   end
 
   def consul_user
@@ -104,6 +108,7 @@ class AccountHelper
       #{mattermost_group}
       #{registry_group}
       #{consul_group}
+      #{prometheus_group}
     )
   end
 end