+++ CHANGELOG.md
@@ -6,6 +6,11 @@ omnibus-gitlab repository.
 11.5.0
 
 - Add option to configure consul group !2781
+- Add Prometheus alerts for GitLab components !2753
+- Update required `gitlab-elasticsearch-indexer` version to 0.3.0 for Elasticsearch 6 support !2791
+- Add option to configure postgresql group !2780
+- Add option to configure prometheus group !2782
+- Add option to configure redis group !2784
 
 11.4.0